Move DoesComponentExist and Log from devfile/adapter to component package
